+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [14.0.0] - 2024-04-09
+
 ### Changed
 
 - [SIL.Archiving] Upgraded to L10nSharp 7.0.0
@@ -379,7 +381,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [SIL.NUnit3Compatibility] new project/package that allows to use NUnit3 syntax with NUnit2
   projects
 
-[Unreleased]: https://github.com/sillsdev/libpalaso/compare/v13.0.1...master
+[Unreleased]: https://github.com/sillsdev/libpalaso/compare/v14.0.0...master
+[14.0.0]: https://github.com/sillsdev/libpalaso/compare/v13.0.1...v14.0.0
 [13.0.1]: https://github.com/sillsdev/libpalaso/compare/v13.0.0...v13.0.1
 [13.0.0]: https://github.com/sillsdev/libpalaso/compare/v12.0.1...v13.0.0
 [12.0.1]: https://github.com/sillsdev/libpalaso/compare/v12.0.0...v12.0.1